SnowsCorrectlySizedButOtherwiseUselessTestOfAnything {TeachingDemos}	R Documentation

Snow's Correctly Sized But Otherwise Useless Test of Anything

Description

This is a hypothesis test designed to be correctly sized in that the probability of rejecting the null when it is true will be equal to your alpha level. Other than that it is a pretty useless test mainly intended for when people say something like "I just need a p-value".

Usage

SnowsCorrectlySizedButOtherwiseUselessTestOfAnything(x,
  data.name = deparse(substitute(x)),
  alternative = "You Are Lucky", ..., seed)

Arguments

x
The data, or nothing, or something equally irrelevant

data.name
The name of the data for the output

alternative
The phrase for the alternate hypothesis in the output

...
Additional arguments that will be silently ignored (like x), future versions may mockingly ignore these instead

seed
A seed (numeric or character) used to seed the random number generator. Use this or manually set the seed if you want reproducible (but still meaningless) results

Details

Some of the advantages/disadvantages of this test include:

The probability of a Type I error is alpha

Power can be easily computed (it is alpha)

Power is independent of the sample size

Power is independent of the hypotheses

This test is not affected by missing data (present data either)

This test does not depend on any distributional or independence assumptions

Value

An object of class htest with the following elements:

p.value
The p-value

statistic
The test statistic (identical to the p-value)

data.name
The name of the data (if any)

method
The name of the test

alternative
a phrase representing the alternative hypothesis

seed
optionally the seed that was used

Note

If someone has suggested that you consider this test, they most likely do not intend for you to actually use the test, rather to reconsider your question or the assumptions that you are making or trying to avoid. This test should only be used to illustrate a point and decisions (other than maybe who should pay for lunch) should never be made based on the results of this test.

Author(s)

Greg Snow 538280@gmail.com

References

The author is unlikely to be willing to publish in any "journal" that would be willing to publish this test.

fortune(264)

See Also

runif

Examples

  SnowsCorrectlySizedButOtherwiseUselessTestOfAnything(log(rnorm(100)))

[Package TeachingDemos version 2.10 Index]