lattice.demo {TeachingDemos}	R Documentation

Interactively explore the conditioned panels in lattice plots.

Description

Plot 1 panel from an xyplot, and optionally a 3d graph highligting the shown points, then allow you to interactively set the conditioning set of data to see the effects and help you better understand how xyplot works.

Usage

lattice.demo(x, y, z, show3d = TRUE)

Arguments

x
The x variable to plot (numeric).

y
The y variable to plot (numeric).

z
The variable to condition on (numeric).

show3d
Logical, should a 3D cloud be shown as well.

Details

This function is intended to for demonstration purposes to help understand what is happening in an xyplot (lattice). When you run the demo it will create a single panel from a conditioned xyplot and optionally a 3D cloud with the points included in the panel highlighted. The function then opens a tcl/tk dialog box that allows you to choose which points are included in the panel (based on the conditioning variable). You can choose the center and width of the shingle displayed and the graph will update to show the new selection.

The intent for this function is for a teacher to show a class how lattice graphics take slices of a 3d plot and show each slice seperately. Students could then work through some examples on their own to better understand what functions like xyplot are doing automatically.

Value

No meaningful return value, this function is run for the side effects.

Author(s)

Greg Snow 538280@gmail.com

See Also

xyplot in lattice package

Examples

if(interactive()){
require(stats)
lattice.demo(quakes$long, quakes$lat, quakes$depth)
}

[Package TeachingDemos version 2.10 Index]