SnowsPenultimateNormalityTest {TeachingDemos}	R Documentation

Test the uninteresting question of whether the data represents an exact normal distribution.

Description

This function tests the null hypothesis that the data comes from an exact normal population. This is a much less interesting/useful null than what people usually want, which is to know if the data come from a distribution that is similar enough to the normal to use normal theory inference.

Usage

SnowsPenultimateNormalityTest(x)

Arguments

x
The data

Details

The theory for this test is based on the probability of getting a rational number from a truly continuous distribution defined on the reals.

The main goal of this test is to quickly give a p-value for those that feel it necessary to test the uninteresting and uninformative null hypothesis that the data represents an exact normal, and allows the user to then move on to much more important questions, like "is the data close enough to the normal to use normal theory inference?".

After running this test (or better instead of running this and any other test of normality) you should ask yourself what it means to test for normality and why you would want to do so. Then plot the data and explore the interesting/useful questions.

Value

An object of class "htest" with components:

p.value
The p-value

alternative
a string representing the alternative hypothesis

method
a string describing the method

data.name
a string describing the name of the data

Note

Note: if you just use this function and report the p-value then the function has failed in its purpose. If this function helps you to think about your analysis and what question(s) you are really interested in, create meaningful plots, and focus on the more meaningful parts of research, then it has succeeded. See also Cochrane's Aphorism.

Author(s)

Greg Snow 538280@gmail.com

References

fortune(234)

See Also

qqnorm, vis.test

Examples

SnowsPenultimateNormalityTest(rt(100,25))

[Package TeachingDemos version 2.10 Index]