power.examp {TeachingDemos}	R Documentation

Graphically illustrate the concept of power.

Description

Create graphs of a normal test statistic under the null and alternative hypotheses to graphically show the idea of power.

Usage

power.examp(n = 1, stdev = 1, diff = 1, alpha = 0.05, xmin = -2, xmax = 4)
run.power.examp(hscale=1.5, vscale=1.5, wait=FALSE)
run.power.examp.old()

Arguments

n
The sample size for the test statistic.

stdev
The standard deviation of the population.

diff
The true difference in means (alternate hypothesis).

alpha
The type I error rate to use for the test.

xmin
The minimum x value to show on the graph.

xmax
The maximum x value to show on the graph.

hscale
Controls width of plot, passed to tkrplot.

vscale
Controls height of plot, passed to tkrplot.

wait
Should R wait for the window to close.

Details

This function will draw 2 graphs representing an upper-tailed test of hypothesis.

The upper panel represents the test statistic under the null hypothesis that the true mean (or mean difference) is 0. It then also shows the upper tail area equal to alpha and the rejection region for the test statistic.

The lower panel shows the normal distribution for the test statistic under the alternative hypothesis where the true mean (or mean difference) is diff. Using the rejection region from the upper panel it shades the upper tail area that corresponds to the power of the test.

Both curves are affected by the specified stdev and sample size n.

The function run.power.examp will in addition create a Tk slider box that will allow you to interactively change the values of stdev, diff, alpha, and n to dynamically see the effects of the change on the graphs and on the power of the test.

This can be used to demonstrate the concept of power, show the effect of sample size on power, show the inverse relationship between the type I and type II error rates, and show how power is dependent on the true mean (or difference) and the population standard deviation.

Value

power.examp invisibly returns the power computed.

run.power.examp returns a list with the parameter settings and the power if wait is TRUE.

run.power.examp.old does not return anything meaningful.

Author(s)

Greg Snow 538280@gmail.com

See Also

power.t.test

Examples

power.examp()
power.examp(n=25)
power.examp(alpha=0.1)

[Package TeachingDemos version 2.10 Index]