run.hist.demo {TeachingDemos}	R Documentation

Create a histogram and interactively change the number of bars.

Description

Create a histogram then use a Tk slider window to change the number of bars, the minimum, and the maximum.

Usage

run.hist.demo(x)

Arguments

x
Data to plot.

Details

Draws a histogram and creates a Tk slider window that allows you to explore how changing the parameters affects the appearance of the plot.

Value

No meaninful value is returned.

Author(s)

Greg Snow 538280@gmail.com

See Also

hist, slider

Examples

if(interactive()){
run.hist.demo( rnorm(250, 100, 5) )
}

[Package TeachingDemos version 2.10 Index]