vis.binom {TeachingDemos}	R Documentation

Plot various distributions then interactivly adjust the parameters.

Description

Plot a curve of a distribution, then using a Tk slider window adjust the parameters and see how the distribution changes. Optionally also plots reference distributions.

Usage

vis.binom()
vis.gamma()
vis.normal()
vis.t()

Details

These functions plot a distribution, then create a Tk slider box that allows you to adjust the parameters of the distribution to see how the curve changes.

Check boxes are available in some cases to also show reference distributions (normal and poisson for the binomial, exponential and chi-squared for gamma, and normal for t). The exponential and chi-squared distributions are those with the same mean as the plotted gamma.

If you change the plotting ranges then you need to click on the 'refresh' button to update the plot.

Value

These functions are run for their side effects and do not return anything meaningful.

Author(s)

Greg Snow 538280@gmail.com

See Also

dnorm, dgamma, etc.

Examples

if(interactive()){
vis.binom()
vis.normal()
vis.gamma()
vis.t()
}

[Package TeachingDemos version 2.10 Index]