z.test {TeachingDemos}	R Documentation

Z test for known population standard deviation

Description

Compute the test of hypothesis and compute confidence interval on the mean of a population when the standard deviation of the population is known.

Usage

z.test(x, mu = 0, stdev, alternative = c("two.sided", "less", "greater"),
  sd = stdev, n=length(x), conf.level = 0.95, ...)

Arguments

x
Vector of data values or the mean of the data.

mu
Hypothesized mean of the population.

stdev
Known standard deviation of the population.

alternative
Direction of the alternative hypothesis.

sd
Alternative to stdev

n
The sample size if x is the sample mean.

conf.level
Confidence level for the interval computation.

...
Additional arguments are silently ignored.

Details

Many introductory statistical texts introduce inference by using the Z test and Z based confidence intervals based on knowing the population standard deviation. Most statistical packages do not include functions to do Z tests since the T test is usually more appropriate for real world situations. This function is meant to be used during that short period of learning when the student is learning about inference using Z procedures, but has not learned the T based procedures yet. Once the student has learned about the T distribution the t.test function should be used instead of this one (but the syntax is very similar, so this function should be an appropriate introductory step to learning t.test).

Value

An object of class htest containing the results

Note

This function should be used for learning only, real data should generally use t.test.

Author(s)

Greg Snow 538280@gmail.com

See Also

t.test, print.htest

Examples

x <- rnorm(25, 100, 5)
z.test(x, 99, 5)

[Package TeachingDemos version 2.10 Index]